split up window.py
